backend/requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.2
certifi==2024.6.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==46.0.3
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==4.2.14
django-cors-headers==4.7.0
django-crispy-forms==2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dpkt==1.9.8
email-validator==2.3.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.2.1
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.2.1
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.3.0
nest-asyncio==1.6.0
oauthlib==3.3.1
openai==1.35.14
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.8.0
pydantic_core==2.20.0
PyGithub==2.8.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.45
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.2.2
virtualenv==20.26.3
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.2.1
zipp==3.19.2
